:wrench: [Chore]: broadcast 및 attendance 엔티티 미사용 column 제거
